pyproject.toml
[tool.poetry]
name = "streamlit-snowflake-analytics"
version = "0.1.0"
description = "A Streamlit application that connects to Snowflake using SSO authentication and runs analytics queries"
authors = ["Your Name <your.email@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.9,<3.9.7 || >3.9.7,<4.0"
streamlit = "^1.32.0"
snowflake-connector-python = "^3.7.0"
pandas = "^2.2.1"
python-dotenv = "^1.0.1"
plotly = "^5.18.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api" 